feat: allow individual queries to set options
